refactor: extract makePassableEncoding
